refactor: finished splitting controllers
