+++ internal/auth/controllers/controller.go
@@ -4,12 +4,10 @@ import (
 	"github.com/indeedhat/barista/internal/auth"
 )
 
-const sessionCookie = "bs"
-
 type Controller struct {
 	repo auth.Repository
 }
 
-func NewController(repo auth.Repository) Controller {
+func New(repo auth.Repository) Controller {
 	return Controller{repo}
 }
