+++ internal/auth/controllers/session_logout.go
@@ -3,12 +3,13 @@ package auth_controllers
 import (
 	"net/http"
 
+	"github.com/indeedhat/barista/internal/auth/cookie"
 	"github.com/indeedhat/barista/internal/ui"
 )
 
 func (c Controller) Logout(rw http.ResponseWriter, r *http.Request) {
 	http.SetCookie(rw, &http.Cookie{
-		Name:     sessionCookie,
+		Name:     cookie.SessionKey,
 		Value:    "",
 		HttpOnly: true,
 		Domain:   r.URL.Host,
