+++ internal/auth/cookie/cookie.go
@@ -1,0 +1,3 @@
+package cookie
+
+const SessionKey = "bs"
