--- internal/brewer/controllers/controllers_basket.go
@@ -1,12 +1,0 @@
-package brewer_controllers
-
-import (
-	"net/http"
-	"strconv"
-
-	"github.com/indeedhat/barista/internal/auth"
-	"github.com/indeedhat/barista/internal/brewer"
-	"github.com/indeedhat/barista/internal/server"
-	"github.com/indeedhat/barista/internal/types"
-	"github.com/indeedhat/barista/internal/ui"
-)
